Refactor visual QA harness outputs
