+++ docs/implementation/visual-qa/transactions/drawer-open-360.png
