+++ docs/implementation/visual-qa/transactions/drawer-open-390.png
