+++ docs/implementation/visual-qa/transactions/expanded-row-1440.png
