+++ docs/implementation/visual-qa/transactions/expanded-row-390.png
